+++ apps/people-app/webapp/src/view/employees/employeeEdit/employeeEdit.tsx
@@ -14,7 +14,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import EmployeeForm from "../EmployeeForm";
+import EmployeeForm from "../onboarding/EmployeeForm";
 
 export default function EmployeeEdit() {
   return <EmployeeForm mode="edit" />;
